Refactor explore screen: Update app bar and progress indicator
Refactor the explore screen to update the app bar and progress indicator. The app bar is now a const widget, and the progress indicator is wrapped in a column to center it vertically. This improves the overall UI and user experience.

Fix date calculation in match table screen

Fix a bug in the match table screen where the selected date was not being adjusted correctly. The date is now correctly adjusted by adding (index - 1) days and then adding 1 day to it. This ensures that the correct date is displayed.

Remove debug print statement in table matches widget

Remove a debug print statement in the table matches widget that was printing the fetched matches date. This statement is no longer needed and has been removed.